+++ pickle-customer/src/main/java/com/example/pickle_customer/order/repository/OrderRepository.java
@@ -1,0 +1,12 @@
+package com.example.pickle_customer.order.repository;
+
+//import com.example.pickle_customer.order.entity.ProductInAccount;
+import com.example.pickle_customer.entity.ProductInAccount;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface OrderRepository extends JpaRepository<ProductInAccount, Long> {
+    List<ProductInAccount> findByProductCode(String productCode);
+
+}
